Dev/S2/dev_Java/dev-oo/src/tp10/PrioritySchedulingQueue.java
package tp10;

public class PrioritySchedulingQueue<T extends IPriority> extends SchedulingQueue<T>
{
    @Override
    public T highestPriority()
    {
        int highPrio = 0;
        for(T ty : theQueue)
        {
            if(ty.getPriority() > highPrio)
            {
                highPrio = ty.getPriority();
            }
        }

        T t = null;
        for(T ty : theQueue)
        {
            if(ty.getPriority() == highPrio)
            {
                t = ty;
            }
        }
        return t;
    }
}
